+++ test/integ_test/aml_model_test/conda.yaml
@@ -9,7 +9,7 @@ dependencies:
       - azure-identity
       - azureml-defaults
       - azureml-dataprep==4.8.4
-      - onnxruntime==1.13.1
+      - onnxruntime
       - datasets
       - scipy
       - transformers
