+++ test/integ_test/aml_resource_path/test_aml_resource_path.py
@@ -2,11 +2,11 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 # --------------------------------------------------------------------------
-import os
 import tempfile
 from pathlib import Path
 
 import pytest
+from test.integ_test.utils import get_olive_workspace_config
 
 from olive.resource_path import ResourceType, create_resource_path
 
@@ -77,25 +77,3 @@ class TestAMLResourcePath:
         saved_resource = resource_path.save_to_dir(self.tmp_dir_path / "save_to_dir", overwrite=True)
         assert Path(saved_resource).exists()
         assert Path(saved_resource).name == "model.onnx"
-
-
-def get_olive_workspace_config():
-    subscription_id = os.environ.get("WORKSPACE_SUBSCRIPTION_ID")
-    if subscription_id is None:
-        raise Exception("Please set the environment variable WORKSPACE_SUBSCRIPTION_ID")
-
-    resource_group = os.environ.get("WORKSPACE_RESOURCE_GROUP")
-    if resource_group is None:
-        raise Exception("Please set the environment variable WORKSPACE_RESOURCE_GROUP")
-
-    workspace_name = os.environ.get("WORKSPACE_NAME")
-    if workspace_name is None:
-        raise Exception("Please set the environment variable WORKSPACE_NAME")
-
-    workspace_config = {
-        "subscription_id": subscription_id,
-        "resource_group": resource_group,
-        "workspace_name": workspace_name,
-    }
-
-    return workspace_config
