+++ test/integ_test/evaluator/azureml_eval/utils.py
@@ -2,11 +2,10 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 # --------------------------------------------------------------------------
-import os
 import shutil
 from pathlib import Path
+from test.integ_test.utils import download_azure_blob, get_olive_workspace_config
 
-from azure.storage.blob import BlobClient
 from torchvision import datasets
 from torchvision.transforms import ToTensor
 
@@ -90,19 +89,6 @@ def get_onnx_model():
     return str(models_dir / "model.onnx")
 
 
-def download_azure_blob(container, blob, download_path):
-    try:
-        conn_str = os.environ["OLIVEWHEELS_STORAGE_CONNECTION_STRING"]
-    except KeyError:
-        raise Exception("Please set the environment variable OLIVEWHEELS_STORAGE_CONNECTION_STRING")
-
-    blob = BlobClient.from_connection_string(conn_str=conn_str, container_name=container, blob_name=blob)
-
-    with open(download_path, "wb") as my_blob:
-        blob_data = blob.download_blob()
-        blob_data.readinto(my_blob)
-
-
 def delete_directories():
     shutil.rmtree(data_dir)
     shutil.rmtree(models_dir)
@@ -123,25 +109,3 @@ def get_aml_target():
         aml_docker_config=docker_config,
         is_dev=True,
     )
-
-
-def get_olive_workspace_config():
-    subscription_id = os.environ.get("WORKSPACE_SUBSCRIPTION_ID")
-    if subscription_id is None:
-        raise Exception("Please set the environment variable WORKSPACE_SUBSCRIPTION_ID")
-
-    resource_group = os.environ.get("WORKSPACE_RESOURCE_GROUP")
-    if resource_group is None:
-        raise Exception("Please set the environment variable WORKSPACE_RESOURCE_GROUP")
-
-    workspace_name = os.environ.get("WORKSPACE_NAME")
-    if workspace_name is None:
-        raise Exception("Please set the environment variable WORKSPACE_NAME")
-
-    workspace_config = {
-        "subscription_id": subscription_id,
-        "resource_group": resource_group,
-        "workspace_name": workspace_name,
-    }
-
-    return workspace_config
